[HWKINVENT-57] Remove the old Tinkerpop impl and replace it with the new one based on the BaseInventory.
