+++ impl-tinkerpop-parent/impl/src/main/java/org/hawkular/inventory/impl/tinkerpop/FilterApplicator.java
@@ -23,20 +23,17 @@ import org.hawkular.inventory.api.filters.Owned;
 import org.hawkular.inventory.api.filters.Related;
 import org.hawkular.inventory.api.filters.RelationWith;
 import org.hawkular.inventory.api.filters.With;
+import org.hawkular.inventory.base.FilterFragment;
+import org.hawkular.inventory.base.Query;
+import org.hawkular.inventory.base.QueryFragment;
 import org.hawkular.inventory.base.spi.SwitchElementType;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Deque;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Consumer;
 
 /**
  * A filter applicator applies a filter to a Gremlin query.
@@ -44,120 +41,18 @@ import java.util.function.Consumer;
  * <p>There is a difference in how certain filters are applied if the query being constructed is considered to be a path
  * to a certain entity or if the filter is used to trim down the number of results.
  *
- * @see PathVisitor
- * @see FilterVisitor
- *
  * @author Lukas Krejci
  * @author Jirka Kremser
+ * @see PathVisitor
+ * @see FilterVisitor
  * @since 0.0.1
  */
 abstract class FilterApplicator<T extends Filter> {
     protected final Type type;
     protected final T filter;
 
-    private FilterApplicator(Type type, T f) {
-        this.type = type;
-        this.filter = f;
-    }
+    private static Map<Class<? extends Filter>, Class<? extends FilterApplicator>> applicators;
 
-    /**
-     * Applies all the filters from the applicator tree to the provided Gremlin query.
-     *
-     * @param filterTree the tree of filters to apply to the query
-     * @param q          the query to update with filters from the tree
-     * @param <S>        type of the source of the query
-     * @param <E>        type of the output of the query
-     */
-    @SuppressWarnings("unchecked")
-    public static <S, E> void applyAll(Tree filterTree, HawkularPipeline<S, E> q) {
-        if (filterTree == null) {
-            return;
-        }
-
-        for (FilterApplicator<?> fa : filterTree.filters) {
-            fa.applyTo(q);
-        }
-
-        if (filterTree.subTrees.isEmpty()) {
-            return;
-        }
-
-        if (filterTree.subTrees.size() == 1) {
-            applyAll(filterTree.subTrees.get(0), q);
-        } else {
-            List<HawkularPipeline<E, ?>> branches = new ArrayList<>();
-            for (Tree t : filterTree.subTrees) {
-                HawkularPipeline<E, ?> branch = new HawkularPipeline<>();
-                applyAll(t, branch);
-                branches.add(branch);
-            }
-
-            // HAWKULAR-255: the dedup pipe in the end could go away once we find out why there are duplicates
-            // TODO: find out what is really happening here
-            q.copySplit(branches.toArray(new HawkularPipeline[branches.size()])).exhaustMerge().dedup();
-        }
-    }
-
-    /**
-     * Converts the list of applicators to the list of filters.
-     *
-     * @param applicators the list of applicators
-     * @return the list of corresponding filters
-     */
-    private static Filter[] filters(FilterApplicator... applicators) {
-        Filter[] ret = new Filter[applicators.length];
-
-        for(int i = 0; i < applicators.length; ++i) {
-            ret[i] = applicators[i].filter;
-        }
-
-        return ret;
-    }
-
-    /**
-     * Converts the provided filter tree to a list of paths (array of filters). Each of the paths in the result
-     * represent 1 branch from the tree from root to some leaf.
-     *
-     * @param applicators the tree of filters
-     * @return the list of paths
-     */
-    public static Filter[][] filters(FilterApplicator.Tree applicators) {
-        List<List<Filter>> paths = new ArrayList<>();
-
-        Deque<Filter[]> path = new ArrayDeque<>();
-        Deque<Iterator<Tree>> traversalPosition = new ArrayDeque<>();
-
-        path.add(filters(applicators.filters));
-        traversalPosition.push(applicators.subTrees.iterator());
-
-        while (!traversalPosition.isEmpty()) {
-            Iterator<Tree> currentPos = traversalPosition.peek();
-            if (currentPos.hasNext()) {
-                Tree child = currentPos.next();
-                if (child.subTrees.isEmpty()) {
-                    //we have a leaf here
-                    List<Filter> pathToLeaf = new ArrayList<>();
-                    for (Filter[] fs : path) {
-                        Collections.addAll(pathToLeaf, fs);
-                    }
-                    Collections.addAll(pathToLeaf, filters(child.filters));
-                    paths.add(pathToLeaf);
-                } else {
-                    path.add(filters(child.filters));
-                    traversalPosition.push(child.subTrees.iterator());
-                }
-            } else {
-                traversalPosition.pop();
-                path.removeLast();
-            }
-        }
-
-        Filter[][] ret = new Filter[paths.size()][];
-        Arrays.setAll(ret, (i) -> paths.get(i).toArray(new Filter[paths.get(i).size()]));
-        return ret;
-    }
-
-    private static Map<Class<? extends  Filter>, Class<? extends FilterApplicator>> applicators;
     static {
         applicators = new HashMap<>();
         applicators.put(Related.class, RelatedApplicator.class);
@@ -171,17 +66,16 @@ abstract class FilterApplicator<T extends Filter> {
         applicators.put(RelationWith.SourceOfType.class, RelationWithSourcesOfTypesApplicator.class);
         applicators.put(RelationWith.TargetOfType.class, RelationWithTargetsOfTypesApplicator.class);
         applicators.put(RelationWith.SourceOrTargetOfType.class, RelationWithSourcesOrTargetsOfTypesApplicator.class);
-        applicators.put(SwitchElementType.class, RelationWithJumpInOutApplicator.class);
+        applicators.put(SwitchElementType.class, SwitchElementTypeApplicator.class);
 
     }
 
-    private static FilterApplicator<?>[] with(Type type, Filter... filters) {
-        FilterApplicator<?>[] ret = new FilterApplicator[filters.length];
-        Arrays.setAll(ret, (i) -> FilterApplicator.with(type, filters[i]));
-        return ret;
+    private FilterApplicator(Type type, T f) {
+        this.type = type;
+        this.filter = f;
     }
 
-    private static FilterApplicator<?> with(Type type, Filter filter) {
+    public static FilterApplicator of(Type type, Filter filter) {
         if (filter == null) {
             throw new IllegalArgumentException("filter == null");
         }
@@ -204,34 +98,49 @@ abstract class FilterApplicator<T extends Filter> {
         try {
             constructor.setAccessible(true);
             return constructor.newInstance(filter, type);
-        } catch(InstantiationException | IllegalAccessException | InvocationTargetException e) {
+        } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
             throw new IllegalArgumentException("Unable to create an instance of " + applicatorClazz);
         }
     }
 
     /**
-     * Creates a new symmetric tree extender that start off from the provided list of paths.
+     * Applies all the filters from the applicator tree to the provided Gremlin query.
      *
-     * @param type    the type of the traversal
-     * @param filters the list of paths
-     * @return an extender that can be used to further extend the provided paths
+     * @param filterTree the tree of filters to apply to the query
+     * @param q          the query to update with filters from the tree
+     * @param <S>        type of the source of the query
+     * @param <E>        type of the output of the query
      */
-    public static SymmetricTreeExtender from(Type type, Filter[][] filters) {
-        return new SymmetricTreeExtender(type, filters);
-    }
+    @SuppressWarnings("unchecked")
+    public static <S, E> void applyAll(Query filterTree, HawkularPipeline<S, E> q) {
+        if (filterTree == null) {
+            return;
+        }
 
-    public static SymmetricTreeExtender fromPath(Filter[][] filters) {
-        return from(Type.PATH, filters);
-    }
+        for (QueryFragment qf : filterTree.getFragments()) {
+            Type applicatorType = (qf instanceof FilterFragment) ? Type.FILTER : Type.PATH;
 
-    public static SymmetricTreeExtender fromPath(Filter... filters) {
-        Filter[][] fs = new Filter[1][];
-        fs[0] = filters;
-        return fromPath(fs);
-    }
+            FilterApplicator.of(applicatorType, qf.getFilter()).applyTo(q);
+        }
 
-    public static SymmetricTreeExtender from(FilterApplicator.Tree filters) {
-        return new SymmetricTreeExtender(filters);
+        if (filterTree.getSubTrees().isEmpty()) {
+            return;
+        }
+
+        if (filterTree.getSubTrees().size() == 1) {
+            applyAll(filterTree.getSubTrees().get(0), q);
+        } else {
+            List<HawkularPipeline<E, ?>> branches = new ArrayList<>();
+            for (Query t : filterTree.getSubTrees()) {
+                HawkularPipeline<E, ?> branch = new HawkularPipeline<>();
+                applyAll(t, branch);
+                branches.add(branch);
+            }
+
+            // HAWKULAR-255: the dedup pipe in the end could go away once we find out why there are duplicates
+            // TODO: find out what is really happening here
+            q.copySplit(branches.toArray(new HawkularPipeline[branches.size()])).exhaustMerge().dedup();
+        }
     }
 
     /**
@@ -262,7 +171,8 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class WithIdsApplicator extends FilterApplicator<With.Ids> {
+    private static final class WithIdsApplicator
+            extends FilterApplicator<With.Ids> {
         private WithIdsApplicator(With.Ids filter, Type type) {
             super(type, filter);
         }
@@ -272,7 +182,8 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class WithTypesApplicator extends FilterApplicator<With.Types> {
+    private static final class WithTypesApplicator
+            extends FilterApplicator<With.Types> {
         private WithTypesApplicator(With.Types filter, Type type) {
             super(type, filter);
         }
@@ -292,7 +203,8 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class RelationWithIdsApplicator extends FilterApplicator<RelationWith.Ids> {
+    private static final class RelationWithIdsApplicator
+            extends FilterApplicator<RelationWith.Ids> {
         private RelationWithIdsApplicator(RelationWith.Ids filter, Type type) {
             super(type, filter);
         }
@@ -302,7 +214,8 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class RelationWithPropertiesApplicator extends FilterApplicator<RelationWith.Properties> {
+    private static final class RelationWithPropertiesApplicator
+            extends FilterApplicator<RelationWith.Properties> {
 
         private RelationWithPropertiesApplicator(RelationWith.Properties filter, Type type) {
             super(type, filter);
@@ -313,7 +226,8 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class RelationWithSourcesOfTypesApplicator extends FilterApplicator<RelationWith
+    private static final class RelationWithSourcesOfTypesApplicator
+            extends FilterApplicator<RelationWith
             .SourceOfType> {
 
         private RelationWithSourcesOfTypesApplicator(RelationWith.SourceOfType filter, Type type) {
@@ -325,8 +239,8 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class RelationWithTargetsOfTypesApplicator extends FilterApplicator<RelationWith
-            .TargetOfType> {
+    private static final class RelationWithTargetsOfTypesApplicator
+            extends FilterApplicator<RelationWith.TargetOfType> {
 
         private RelationWithTargetsOfTypesApplicator(RelationWith.TargetOfType filter, Type type) {
             super(type, filter);
@@ -338,7 +252,8 @@ abstract class FilterApplicator<T extends Filter> {
 
     }
 
-    private static final class RelationWithSourcesOrTargetsOfTypesApplicator extends FilterApplicator<RelationWith
+    private static final class RelationWithSourcesOrTargetsOfTypesApplicator
+            extends FilterApplicator<RelationWith
             .SourceOrTargetOfType> {
         private RelationWithSourcesOrTargetsOfTypesApplicator(RelationWith.SourceOrTargetOfType filter, Type type) {
             super(type, filter);
@@ -349,168 +264,14 @@ abstract class FilterApplicator<T extends Filter> {
         }
     }
 
-    private static final class RelationWithJumpInOutApplicator extends FilterApplicator<SwitchElementType> {
-        private RelationWithJumpInOutApplicator(SwitchElementType filter, Type type) {
+    private static final class SwitchElementTypeApplicator
+            extends FilterApplicator<SwitchElementType> {
+        private SwitchElementTypeApplicator(SwitchElementType filter, Type type) {
             super(type, filter);
         }
 
         public void applyTo(HawkularPipeline<?, ?> query) {
             type.visitor.visit(query, filter);
-        }
-    }
-
-    /**
-     * Constructs an filter applicator tree by extending all the leaves with a uniform set of filters at a time.
-     */
-    static class SymmetricTreeExtender {
-        private Tree.Builder filters;
-
-        private SymmetricTreeExtender(Type type, Filter[][] filters) {
-            this.filters = new Tree.Builder();
-            and(type, filters);
-        }
-
-        private SymmetricTreeExtender(FilterApplicator.Tree filters) {
-            this.filters = filters.asBuilder();
-        }
-
-        public SymmetricTreeExtender and(Type type, Filter[][] filters) {
-            onLeaves(this.filters, (t) -> {
-                for(Filter[] fs : filters) {
-                    t.branch().with(FilterApplicator.with(type, fs));
-                }
-            });
-            return this;
-        }
-
-        public SymmetricTreeExtender and(Type type, Filter... filters) {
-            onLeaves(this.filters, (t) -> t.with(FilterApplicator.with(type, filters)));
-            return this;
-        }
-
-        public SymmetricTreeExtender andFilter(Filter... filters) {
-            return and(Type.FILTER, filters);
-        }
-
-        public SymmetricTreeExtender andPath(Filter... path) {
-            return and(Type.PATH, path);
-        }
-
-        public SymmetricTreeExtender andPath(Filter[][] path) {
-            return and(Type.PATH, path);
-        }
-
-        public FilterApplicator.Tree get() {
-            return filters.build();
-        }
-
-        private void onLeaves(Tree.Builder root, Consumer<Tree.Builder> leafMutator) {
-            if (root.children.isEmpty()) {
-                leafMutator.accept(root);
-            } else {
-                for(Tree.Builder c : root.children) {
-                    onLeaves(c, leafMutator);
-                }
-            }
-        }
-    }
-
-    /**
-     * Represents a tree of filter applicators.
-     * The applicators that are used to represent a Gremlin query (so that new query instances can be generated at will)
-     * form a tree because of the fact, environments can both contain resources or metrics directly or they can contain
-     * feeds that in turn contain the metrics and resources.
-     *
-     * <p>The API contains a method to traverse both of these kinds of "placement" of metrics or resources
-     * simultaneously. This is nice from a user perspective but creates problems in the impl, because using the
-     * relationships of the resources or metrics, one can construct queries, the would be able to branch multiple times
-     * and hence form a tree.
-     */
-    public static final class Tree {
-        FilterApplicator<?>[] filters;
-        List<Tree> subTrees = new ArrayList<>();
-
-        private Tree() {}
-
-        public Builder asBuilder() {
-            Builder b = new Builder();
-            b.filters = new ArrayList<>(Arrays.asList(filters));
-            for (Tree subTree : subTrees) {
-                Builder childBuilder = subTree.asBuilder();
-                childBuilder.parent = b;
-                b.children.add(childBuilder);
-            }
-
-            return b;
-        }
-
-        public static final class Builder {
-            List<FilterApplicator<?>> filters = new ArrayList<>();
-            Tree tree = new Tree();
-            Builder parent;
-            List<Builder> children = new ArrayList<>();
-            boolean done;
-
-            /**
-             * Creates a new branch in the tree and returns a builder of that branch.
-             */
-            Builder branch() {
-                Builder child = new Builder();
-                child.parent = this;
-                children.add(child);
-
-                return child;
-            }
-
-            /**
-             * Concludes the work on a branch and returns a builder of the parent "node", if any.
-             */
-            Builder done() {
-                if (done) {
-                    return parent;
-                }
-
-                this.tree.filters = filters.toArray(new FilterApplicator[filters.size()]);
-                if (parent != null) {
-                    parent.tree.subTrees.add(this.tree);
-                    parent.children.remove(this);
-                }
-
-                //done will remove the child from the children, so we'd get concurrent modification exceptions
-                //avoid that stupidly by working on a copy of children
-                new ArrayList<>(children).forEach(Tree.Builder::done);
-
-                done = true;
-
-                return parent;
-            }
-
-            /**
-             * Sets the filters to be used on the current node in the tree.
-             * @param filters the list of filters to apply to the query at this position in the tree.
-             * @return
-             */
-            Builder with(FilterApplicator<?>... filters) {
-                Collections.addAll(this.filters, filters);
-                return this;
-            }
-
-            /**
-             * Builds the <b>whole</b> tree regardless of where in the tree the current builder "operates".
-             * @return the fully built tree
-             */
-            Tree build() {
-                Tree.Builder root = this;
-                while (true) {
-                    if (root.parent == null) {
-                        break;
-                    }
-                    root = root.parent;
-                }
-
-                root.done();
-                return root.tree;
-            }
         }
     }
 }
