+++ impl-tinkerpop-parent/impl/src/main/java/org/hawkular/inventory/impl/tinkerpop/HawkularPipeline.java
@@ -121,15 +121,15 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
         return in(srels);
     }
 
-    public HawkularPipeline<S, Vertex> page(Pager pager) {
-        return cast(Vertex.class).page(pager, (e, p) -> {
+    public HawkularPipeline<S, ? extends Element> page(Pager pager) {
+        return cast(Element.class).page(pager, (e, p) -> {
             String prop = Constants.Property.mapUserDefined(p);
             return e.getProperty(prop);
         });
     }
 
-    public <V extends Comparable<V>> HawkularPipeline<S, E> page(Pager pager,
-            BiFunction<E, String, V> propertyValueExtractor) {
+    public HawkularPipeline<S, E> page(Pager pager,
+            BiFunction<E, String, ? extends Comparable> propertyValueExtractor) {
 
         List<Order> order = pager.getOrder();
         if (!order.isEmpty()) {
@@ -148,8 +148,8 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
                     int ret = 0;
                     for (Order ord : order) {
                         if (ord.isSpecific()) {
-                            V a = propertyValueExtractor.apply(p.getA(), ord.getField());
-                            V b = propertyValueExtractor.apply(p.getB(), ord.getField());
+                            Comparable a = propertyValueExtractor.apply(p.getA(), ord.getField());
+                            Comparable b = propertyValueExtractor.apply(p.getB(), ord.getField());
                             ret = ord.isAscending() ? safeCompare(a, b) : safeCompare(b, a);
                             if (ret != 0) {
                                 break;
@@ -358,7 +358,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
 
     @Override
     public HawkularPipeline<S, E> groupBy(PipeFunction keyFunction, PipeFunction valueFunction,
-                                          PipeFunction reduceFunction) {
+            PipeFunction reduceFunction) {
         return cast(super.groupBy(keyFunction, valueFunction, reduceFunction));
     }
 
@@ -369,7 +369,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
 
     @Override
     public HawkularPipeline<S, E> groupBy(Map reduceMap, PipeFunction keyFunction, PipeFunction valueFunction,
-                                          PipeFunction reduceFunction) {
+            PipeFunction reduceFunction) {
         return cast(super.groupBy(reduceMap, keyFunction, valueFunction, reduceFunction));
     }
 
@@ -385,7 +385,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
 
     @Override
     public HawkularPipeline<S, E> groupCount(PipeFunction keyFunction,
-                                             PipeFunction<Pair<?, Number>, Number> valueFunction) {
+            PipeFunction<Pair<?, Number>, Number> valueFunction) {
         return cast(super.groupCount(keyFunction, valueFunction));
     }
 
@@ -401,7 +401,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
 
     @Override
     public HawkularPipeline<S, E> groupCount(Map<?, Number> map, PipeFunction keyFunction,
-                                             PipeFunction<Pair<?, Number>, Number> valueFunction) {
+            PipeFunction<Pair<?, Number>, Number> valueFunction) {
         return cast(super.groupCount(map, keyFunction, valueFunction));
     }
 
@@ -452,7 +452,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
 
     @Override
     public HawkularPipeline<S, ?> ifThenElse(PipeFunction<E, Boolean> ifFunction, PipeFunction<E, ?> thenFunction,
-                                             PipeFunction<E, ?> elseFunction) {
+            PipeFunction<E, ?> elseFunction) {
         return cast(super.ifThenElse(ifFunction, thenFunction, elseFunction));
     }
 
@@ -533,7 +533,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
 
     @Override
     public HawkularPipeline<S, E> loop(String namedStep, PipeFunction<LoopPipe.LoopBundle<E>, Boolean> whileFunction,
-                                       PipeFunction<LoopPipe.LoopBundle<E>, Boolean> emitFunction) {
+            PipeFunction<LoopPipe.LoopBundle<E>, Boolean> emitFunction) {
         return cast(super.loop(namedStep, whileFunction, emitFunction));
     }
 
@@ -546,7 +546,7 @@ final class HawkularPipeline<S, E> extends GremlinPipeline<S, E> implements Clon
     @Override
     @Deprecated
     public HawkularPipeline<S, E> loop(int numberedStep, PipeFunction<LoopPipe.LoopBundle<E>, Boolean> whileFunction,
-                                       PipeFunction<LoopPipe.LoopBundle<E>, Boolean> emitFunction) {
+            PipeFunction<LoopPipe.LoopBundle<E>, Boolean> emitFunction) {
         return cast(super.loop(numberedStep, whileFunction, emitFunction));
     }
