+++ impl-tinkerpop-parent/impl/src/main/resources/META-INF/services/org.hawkular.inventory.api.Inventory
@@ -1,1 +1,1 @@
-org.hawkular.inventory.impl.tinkerpop.InventoryService
+org.hawkular.inventory.impl.tinkerpop.TinkerpopInventory
