+++ impl-tinkerpop-parent/impl/src/test/java/org/hawkular/inventory/impl/tinkerpop/test/TinkerpopTest.java
@@ -19,7 +19,7 @@ package org.hawkular.inventory.impl.tinkerpop.test;
 import com.tinkerpop.blueprints.Element;
 import org.hawkular.inventory.api.test.AbstractBaseInventoryPersistenceCheck;
 import org.hawkular.inventory.base.BaseInventory;
-import org.hawkular.inventory.impl.tinkerpop.lazy.TinkerpopInventory;
+import org.hawkular.inventory.impl.tinkerpop.TinkerpopInventory;
 
 import java.io.IOException;
 import java.nio.file.FileVisitResult;
@@ -33,7 +33,7 @@ import java.nio.file.attribute.BasicFileAttributes;
  * @author Lukas Krejci
  * @since 0.0.6
  */
-public class BaseTest extends AbstractBaseInventoryPersistenceCheck<Element> {
+public class TinkerpopTest extends AbstractBaseInventoryPersistenceCheck<Element> {
     @Override
     protected BaseInventory<Element> instantiateNewInventory() {
         return new TinkerpopInventory();
